refactoring, extracted SanityMath
