+++ Assets/Scripts/Entity/Player/Sanity.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 029f06f580ce65f47b73af332b4e6b08
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
