+++ Assets/Scripts/Entity/Player/Sanity/SanityMath.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 00698a7b54a6db844b24496de4a13691
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
